# openapi.yaml
openapi: 3.0.0
info:
  title: GDG Subscription API
  description: A production-ready API for subscription-based content delivery.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /admin/logs:
    get:
      summary: Get system logs
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
          description: Number of logs to return
      responses:
        '200':
          description: Logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/report/csv:
    get:
      summary: Export system report as CSV
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: CSV report generated
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users:
    get:
      summary: Get list of users
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
          nullable: true
          example: null
        message:
          type: string
          example: "An error occurred"
        timestamp:
          type: string
          format: date-time
          example: "2023-10-25T12:00:00Z"
...
    LogsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: "log-123"
              level:
                type: string
                example: "INFO"
              message:
                type: string
                example: "User logged in"
              createdAt:
                type: string
                format: date-time
                example: "2023-10-25T12:00:00Z"
        message:
          type: string
          example: "Logs retrieved successfully"
        timestamp:
          type: string
          format: date-time
          example: "2023-10-25T12:00:00Z"
    UsersListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: "uuid-1234"
              email:
                type: string
                example: "user@example.com"
              name:
                type: string
                example: "John Doe"
              role:
                type: string
                example: "USER"
              tier:
                type: string
                example: "PREMIUM"
        message:
          type: string
          example: "Users retrieved successfully"
        timestamp:
          type: string
          format: date-time
          example: "2023-10-25T12:00:00Z"
...